+++ microbit-common/src/v2/board.rs
@@ -134,6 +134,8 @@ impl Board {
     ///
     /// This method will return an instance of the board the first time it is
     /// called. It will return only `None` on subsequent calls.
+    /// This function can also return `None` if one of the the peripherals was
+    /// already taken.
     pub fn take() -> Option<Self> {
         Some(Self::new(
             pac::Peripherals::take()?,
@@ -141,7 +143,14 @@ impl Board {
         ))
     }
 
-    fn new(p: pac::Peripherals, cp: pac::CorePeripherals) -> Self {
+    /// Fallback method in the case peripherals and core peripherals were taken
+    /// elsewhere already.
+    ///
+    /// This method will take the peripherals and core peripherals and
+    /// return an instance of the board.
+    ///
+    /// An exemplary usecase is shown in the rtic display example.
+    pub fn new(p: pac::Peripherals, cp: pac::CorePeripherals) -> Self {
         let p0parts = p0::Parts::new(p.P0);
         let p1parts = p1::Parts::new(p.P1);
         Self {