+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- Make `Board::new(p, cp)` public and fix RTIC example
 - Fix display-nonblocking example
 
 ## [0.12.0] - 2021-11-10